A lot of unstable changes
Quick note, this version is unstable, as the work done is integrating a
lot of new features (SSL decentralisation) so do not expect to run full
on your machine
However currently SSL has been achieved a part of the connection
packages, also the SSL connections to the SQL servers work. The sql ID
servers can now be setup
Another note there is a datastructure for the ID servers set up, there
is more I can note but I cannot be bothered to note them all.