Ok it's 3am and here is a list of what is done
Been able to partial setup the listening of three ports all at the same
time, need to adjust it a small bit to make it more effective.
The encrpytion of the secure details is done and from what I can tell
work, I need to fully test it.
The SQL databases are setup and the details are stored in a stack. Going
to fully stress test the server will tell you when I have crashed the
entire comsci network :)